gui: Add header file for x_newtext.c
